server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
firebase-admin==6.6.0
google-generativeai==0.8.3
pytz==2024.2
python-dotenv==1.0.1